+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix cache resolution for alternative networks [#151]
+- Fix cache error detection [#163]
 
 ### Removed
 
@@ -370,6 +371,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implementation of `Store` trait from `wallet-core`
 - Implementation of `State` and `Prover` traits from `wallet-core`
 
+[#163]: https://github.com/dusk-network/wallet-cli/issues/163
 [#151]: https://github.com/dusk-network/wallet-cli/issues/151
 [#144]: https://github.com/dusk-network/wallet-cli/issues/144
 [#133]: https://github.com/dusk-network/wallet-cli/issues/133